+++ loader.js
@@ -1,6 +1,0 @@
-const { Schematic, SchematicHelpers } = require('./src/schematic.js');
-
-module.exports = {
-  Schematic: Schematic,
-  app: new SchematicHelpers(),
-};